# swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Simple Shop API",
        "description": "A simple shop API",
        "version": "1.0.0"
    },
    "servers": [
        {
          "url": "http://localhost:8080/api/v1",
          "description": "Local server"
        }
    ],
    "paths": {
...
        "/orders": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": ["Order operations"],
                "description": "Returns a list of users's orders",
                "operationId": "findAllOrders",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Order"
                                    }                                   
                                }                                
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid token",
                        "content": {
                            "application/json": {
                                "schema": {  
                                    "$ref": "#/components/schemas/Error"                                  
                                }                                
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {  
                                    "$ref": "#/components/schemas/Error"                                  
                                }                              
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": ["Order operations"],
                "description": "Creates a new order",
                "operationId": "createOrder",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {                               
                                "$ref": "#/components/schemas/ShortOrderData"                                
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "New order created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }                                
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid token",
                        "content": {
                            "application/json": {
                                "schema": {  
                                    "$ref": "#/components/schemas/Error"                                  
                                }                                
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {  
                                    "$ref": "#/components/schemas/Error"                                  
                                }                              
                            }
                        }
                    }
                }
            }
        },
        "/orders/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": ["Order operations"],
                "description": "Returns order by id",
                "operationId": "findOrderById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true 
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }                                
                            }
                        }
                    },
                    "404": {
                        "description": "Order is not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                } 
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid token",
                        "content": {
                            "application/json": {
                                "schema": {  
                                    "$ref": "#/components/schemas/Error"                                  
                                }                                
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {  
                                    "$ref": "#/components/schemas/Error"                                  
                                }                              
                            }
                        }
                    }
                } 
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": ["Order operations"],
                "description": "Updates an existing order",
                "operationId": "updateOrder",
                "parameters": [
                    {                           
                        "name": "id",
                        "in": "path",
                        "required": true                    
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {                               
                                "$ref": "#/components/schemas/ShortOrderData"                                
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "An existing order updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }                                
                            }
                        }   
                    },
                    "404": {
                        "description": "Order is not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                } 
                            }
                        }    
                    },
                    "401": {
                        "description": "Invalid token",
                        "content": {
                            "application/json": {
                                "schema": {  
                                    "$ref": "#/components/schemas/Error"                                  
                                }                                
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {  
                                    "$ref": "#/components/schemas/Error"                                  
                                }                              
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": ["Order operations"],
                "description": "Deletes an existing order",
                "operationId": "deleteOrder",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true 
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An existing order deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }                             
                            }
                        }   
                    },
                    "404": {
                        "description": "Order is not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                } 
                            }
                        }    
                    },
                    "401": {
                        "description": "Invalid token",
                        "content": {
                            "application/json": {
                                "schema": {  
                                    "$ref": "#/components/schemas/Error"                                  
                                }                                
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {  
                                    "$ref": "#/components/schemas/Error"                                  
                                }                              
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "ShortProductData": {
                "type": "object",
                "properties": {
                    "_id": {
                        "type": "string"
                    },
                    "quantity": {
                        "type": "number"
                    }
                },
                "required": ["_id", "quantity"]
            },
            "Order": {
                "type": "object",
                "properties": {
                    "_id": {
                        "type": "string"
                    },
                    "userId": {
                        "type": "string"
                    },
                    "products": {
                        "type": "array",   
                        "items": {
                            "$ref": "#/components/schemas/ShortProductData"
                        }                                   
                    },
                    "totalPrice": {
                        "type": "number"
                    }
                }
            },
            "ShortOrderData": {
                "type": "object",
                "properties": {
                    "products": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ShortProductData"
                        }   
                    }
                }
            },
            "Error": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "statusCode": {
                        "type": "string"
                    }
                }
            }    
        },
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}